Move all tests from Rackspace to OpenStack.
