+++ test/compute/test_rackspace.py
@@ -15,11 +15,8 @@
 import sys
 import unittest
 
-from libcloud.common.types import InvalidCredsError, MalformedResponseError
-from libcloud.compute.drivers.rackspace import RackspaceNodeDriver as Rackspace
-from libcloud.compute.base import Node, NodeImage, NodeSize
-
-from test.compute.test_openstack import OpenStackTests, OpenStackMockHttp
+from libcloud.compute.drivers.rackspace import RackspaceNodeDriver
+from test.compute.test_openstack import OpenStackTests
 
 from test.secrets import RACKSPACE_USER, RACKSPACE_KEY
 
@@ -28,212 +25,11 @@ class RackspaceTests(OpenStackTests):
     should_list_locations = True
     should_have_pricing = True
 
-    driver_type = Rackspace
+    driver_type = RackspaceNodeDriver
     driver_args = (
         RACKSPACE_USER,
         RACKSPACE_KEY,
     )
-
-    def test_auth(self):
-        OpenStackMockHttp.type = 'UNAUTHORIZED'
-        try:
-            self.driver = self.create_driver()
-        except InvalidCredsError, e:
-            self.assertEqual(True, isinstance(e, InvalidCredsError))
-        else:
-            self.fail('test should have thrown')
-
-    def test_auth_missing_key(self):
-        OpenStackMockHttp.type = 'UNAUTHORIZED_MISSING_KEY'
-        try:
-            self.driver = self.create_driver()
-        except MalformedResponseError, e:
-            self.assertEqual(True, isinstance(e, MalformedResponseError))
-        else:
-            self.fail('test should have thrown')
-
-    def test_auth_server_error(self):
-        OpenStackMockHttp.type = 'INTERNAL_SERVER_ERROR'
-        try:
-            self.driver = self.create_driver()
-        except MalformedResponseError, e:
-            self.assertEqual(True, isinstance(e, MalformedResponseError))
-        else:
-            self.fail('test should have thrown')
-
-    def test_list_nodes(self):
-        OpenStackMockHttp.type = 'EMPTY'
-        ret = self.driver.list_nodes()
-        self.assertEqual(len(ret), 0)
-        OpenStackMockHttp.type = None
-        ret = self.driver.list_nodes()
-        self.assertEqual(len(ret), 1)
-        node = ret[0]
-        self.assertEqual('67.23.21.33', node.public_ip[0])
-        self.assertEqual('10.176.168.218', node.private_ip[0])
-        self.assertEqual(node.extra.get('flavorId'), '1')
-        self.assertEqual(node.extra.get('imageId'), '11')
-        self.assertEqual(type(node.extra.get('metadata')), type(dict()))
-        OpenStackMockHttp.type = 'METADATA'
-        ret = self.driver.list_nodes()
-        self.assertEqual(len(ret), 1)
-        node = ret[0]
-        self.assertEqual(type(node.extra.get('metadata')), type(dict()))
-        self.assertEqual(node.extra.get('metadata').get('somekey'),
-                         'somevalue')
-        OpenStackMockHttp.type = None
-
-    def test_list_sizes(self):
-        ret = self.driver.list_sizes()
-        self.assertEqual(len(ret), 7)
-        size = ret[0]
-        self.assertEqual(size.name, '256 slice')
-        self.assertTrue(isinstance(size.price, float))
-
-    def test_list_images(self):
-        ret = self.driver.list_images()
-        expected = {10: {'serverId': None,
-                         'status': 'ACTIVE',
-                         'created': '2009-07-20T09:14:37-05:00',
-                         'updated': '2009-07-20T09:14:37-05:00',
-                         'progress': None},
-                    11: {'serverId': '91221',
-                         'status': 'ACTIVE',
-                         'created': '2009-11-29T20:22:09-06:00',
-                         'updated': '2009-11-29T20:24:08-06:00',
-                         'progress': '100'}}
-        for ret_idx, extra in expected.items():
-            for key, value in extra.items():
-                self.assertEqual(ret[ret_idx].extra[key], value)
-
-    def test_create_node(self):
-        image = NodeImage(id=11, name='Ubuntu 8.10 (intrepid)',
-                          driver=self.driver)
-        size = NodeSize(1, '256 slice', None, None, None, None,
-                        driver=self.driver)
-        node = self.driver.create_node(name='racktest', image=image, size=size)
-        self.assertEqual(node.name, 'racktest')
-        self.assertEqual(node.extra.get('password'), 'racktestvJq7d3')
-
-    def test_create_node_ex_shared_ip_group(self):
-        OpenStackMockHttp.type = 'EX_SHARED_IP_GROUP'
-        image = NodeImage(id=11, name='Ubuntu 8.10 (intrepid)',
-                          driver=self.driver)
-        size = NodeSize(1, '256 slice', None, None, None, None,
-                        driver=self.driver)
-        node = self.driver.create_node(name='racktest', image=image, size=size,
-                                       ex_shared_ip_group_id='12345')
-        self.assertEqual(node.name, 'racktest')
-        self.assertEqual(node.extra.get('password'), 'racktestvJq7d3')
-
-    def test_create_node_with_metadata(self):
-        OpenStackMockHttp.type = 'METADATA'
-        image = NodeImage(id=11, name='Ubuntu 8.10 (intrepid)',
-                          driver=self.driver)
-        size = NodeSize(1, '256 slice', None, None, None, None,
-                        driver=self.driver)
-        metadata = {'a': 'b', 'c': 'd'}
-        files = {'/file1': 'content1', '/file2': 'content2'}
-        node = self.driver.create_node(name='racktest', image=image, size=size,
-                                       metadata=metadata, files=files)
-        self.assertEqual(node.name, 'racktest')
-        self.assertEqual(node.extra.get('password'), 'racktestvJq7d3')
-        self.assertEqual(node.extra.get('metadata'), metadata)
-
-    def test_reboot_node(self):
-        node = Node(id=72258, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        ret = node.reboot()
-        self.assertTrue(ret is True)
-
-    def test_destroy_node(self):
-        node = Node(id=72258, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        ret = node.destroy()
-        self.assertTrue(ret is True)
-
-    def test_ex_limits(self):
-        limits = self.driver.ex_limits()
-        self.assertTrue("rate" in limits)
-        self.assertTrue("absolute" in limits)
-
-    def test_ex_save_image(self):
-        node = Node(id=444222, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        image = self.driver.ex_save_image(node, "imgtest")
-        self.assertEqual(image.name, "imgtest")
-        self.assertEqual(image.id, "12345")
-
-    def test_ex_list_ip_addresses(self):
-        ret = self.driver.ex_list_ip_addresses(node_id=72258)
-        self.assertEquals(2, len(ret.public_addresses))
-        self.assertTrue('67.23.10.131' in ret.public_addresses)
-        self.assertTrue('67.23.10.132' in ret.public_addresses)
-        self.assertEquals(1, len(ret.private_addresses))
-        self.assertTrue('10.176.42.16' in ret.private_addresses)
-
-    def test_ex_list_ip_groups(self):
-        ret = self.driver.ex_list_ip_groups()
-        self.assertEquals(2, len(ret))
-        self.assertEquals('1234', ret[0].id)
-        self.assertEquals('Shared IP Group 1', ret[0].name)
-        self.assertEquals('5678', ret[1].id)
-        self.assertEquals('Shared IP Group 2', ret[1].name)
-        self.assertTrue(ret[0].servers is None)
-
-    def test_ex_list_ip_groups_detail(self):
-        ret = self.driver.ex_list_ip_groups(details=True)
-
-        self.assertEquals(2, len(ret))
-
-        self.assertEquals('1234', ret[0].id)
-        self.assertEquals('Shared IP Group 1', ret[0].name)
-        self.assertEquals(2, len(ret[0].servers))
-        self.assertEquals('422', ret[0].servers[0])
-        self.assertEquals('3445', ret[0].servers[1])
-
-        self.assertEquals('5678', ret[1].id)
-        self.assertEquals('Shared IP Group 2', ret[1].name)
-        self.assertEquals(3, len(ret[1].servers))
-        self.assertEquals('23203', ret[1].servers[0])
-        self.assertEquals('2456', ret[1].servers[1])
-        self.assertEquals('9891', ret[1].servers[2])
-
-    def test_ex_create_ip_group(self):
-        ret = self.driver.ex_create_ip_group('Shared IP Group 1', '5467')
-        self.assertEquals('1234', ret.id)
-        self.assertEquals('Shared IP Group 1', ret.name)
-        self.assertEquals(1, len(ret.servers))
-        self.assertEquals('422', ret.servers[0])
-
-    def test_ex_delete_ip_group(self):
-        ret = self.driver.ex_delete_ip_group('5467')
-        self.assertEquals(True, ret)
-
-    def test_ex_share_ip(self):
-        ret = self.driver.ex_share_ip('1234', '3445', '67.23.21.133')
-        self.assertEquals(True, ret)
-
-    def test_ex_unshare_ip(self):
-        ret = self.driver.ex_unshare_ip('3445', '67.23.21.133')
-        self.assertEquals(True, ret)
-
-    def test_ex_resize(self):
-        node = Node(id=444222, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        size = NodeSize(1, '256 slice', None, None, None, None,
-                        driver=self.driver)
-        self.assertTrue(self.driver.ex_resize(node=node, size=size))
-
-    def test_ex_confirm_resize(self):
-        node = Node(id=444222, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        self.assertTrue(self.driver.ex_confirm_resize(node=node))
-
-    def test_ex_revert_resize(self):
-        node = Node(id=444222, name=None, state=None, public_ip=None,
-                    private_ip=None, driver=self.driver)
-        self.assertTrue(self.driver.ex_revert_resize(node=node))
 
 
 if __name__ == '__main__':
